[FEAT] BretteSoft Géronimo
if return exceed 1 month and two weeks, ban user for one month (cheh)

+++ src/App/Server/Timers/ITimer.java
@@ -3,6 +3,6 @@ package App.Server.Timers;
 public interface ITimer extends Runnable {
 
     String getName();
-    int getDuration();
+    long getDuration();
 
 }
